To streamline testing and prevent missed errors, I updated README.md with instructions to run 'test' as part of the spring-boot startup.
Updated unit tests to handle some changed logic and removed one which is weirdly broken, still not sure why, but I'm going to change the logic on how the services handle bad input. Also, fixed a warning about Mockito needing an agent configuration in the near future (past Java 21) by changing the pom file and adding a pair of JVM settings. The first one was suggested ("-XX:+EnableDynamicAgentLoading"), the another warning appeared after setting that one but not as urgent ("-Xshare:off"). Might be swalling the spider to catch the fly... however, so far no tummy ache.